temporarily disabled opt mode for Intel oneAPI DPC++ compilers
There seems to be a serious optimization bug in the new Intel
C++ compilers, since it consistently produces incorrect results
related to the computation of our jacobian determinants. which
are used throughout the FEM kernels. Adding compiler hacks for
this problem does not make any sense, so for now, opt-mode is
disabled for ICPX and we'll see how Intel reacts to our bug
report, see:

https://community.intel.com/t5/Intel-oneAPI-DPC-C-Compiler/Optimization-compiler-bug-in-Intel-DPC-compiler-2024-and-older/m-p/1617029#

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>