replace null rows in AmaVanka by unit rows in case of singular macros
In the case where an AmaVanka smoother is created on a mesh that contains only
a single element with no explicitly incorporated boundary conditions, the local
Stokes matrix on that single element is singular and therefore the local
inversion fails; this can happen if there are only slip boundary conditions
on the entire boundary.

Up to now, the AmaVanka matrix would have simply been a null matrix and
therefore the preconditioner would become singular. This commit changes this
behaviour in the sense that null rows are now replaced by unit rows, so that
the AmaVanka matrix becomes the identity matrix in this case.

This commit also adds a bew check to IterativeSolver::_set_initial_defect,
which checks the defect against _tol_abs_low and returns Status::success in
this case.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>