removed float cast in CUDA component_product kernel and fixed some tests
The broken kernel was not uncovered by the unit test system, because the
tolerances in the test were chosen far too low as eps^(0.4), which evaluates
to rougly 4e-7 for double precision, which is lower than the machine precision
of float -- and this is why the double-to-float cast was not uncovered.

As a general guideline for choosing the tolerances in unit tests, one should
try to choose something like eps^(0.8) first, where eps is the machine precision
of the selected floating point format; this should be good enough for most tested
operations. If the test fails this tolerance, further investigations have to
be made to figure out whether the tested code is broken or whether the compiler
has found an aggressive optimization strategy that leads to lower precision
results.

There is one notable exception: If the tested operation contains square-roots
or similar operations, then the tolerance has to be set to a fairly small value,
because computing the square root will get merely the first half digits correctly
in a finite precision world. For example, when computing the euclidean norm, the
tolerance has to be chosen as something like eps^(0.4) -- or one might choose to
test the squared result against the standard tolerance instead.

In any case, do not simply lower the tolerances until the tests pass without
actually checking whether the results make sense, because this actually renders
the whole concept of unit tests ad absurdum.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>