added steady DFG95 CCND hypbrid HPC benchmark application
This commit adds yet-another-dfg95-ccnd-application (TM) to FEAT3, which is a
direct offspring of the existing "dfg95-ccnd-steady" application. This new
application supports multi-precision solvers, multi-threading and will
be extended to support CUDA-based assembly and solvers in near future.

It is not recommended to use this code as a base for own CCND-style applications,
because this code is hard-wired to solve the DFG95 benchmarks and is meant as a
testing and proving ground for our hybrid-high-performance-computing efforts,
which requires a certain minimum amount of care to use due to its multi-precision
implementation.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>