some minor improvements of iso-parametric trafo implementation
The primary change of this commit is the addittion of the add_trafo_mesh_part_charts
function to the DomainControl, which loops over all mesh parts on all mesh levels
and adds those which are attached to a chart to the transformation.

Additionally, this commit introduces two new analytic functions:

* Analytic::Distance::InverseDistanceFunction
This class implements u(x) := 1/|x-x_0|_2 for a given origin point x_0.

* Analytic::Common::HarmonicShellFunction
This class implements a harmonic function that is well defined in all points
except for x=x_0. This function is equal to the Green's function for the
Laplace operator, scaled such that u(x) = 1 for |x - x_0|_2 = 1 and u(x) = 2
for |x - x_0|_2 = 1/2, and can therefore be used as an analytical function
for the Laplace equation defined on a 2D ring or 3D spherical shell, i.e.
a ball with a ball cut out.

Finally, the hessian matrix computation of the 3D hexahedral Q3 element has been
added.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>