improved and corrected documentation of inner residual scale for FGMRES solver
The previous documentation was incorrect in claiming that the inner residuals
were unrelated to the actual residual norms in case of preconditiong. Mea culpa.

Here's an excerpt from the new corrected documentation:

In contrast to other popular iterative Krylov subspace methods like PCG or BiCGStab, the GMRES method and all
of its variants -- including the FMGRES(k) solver implemented by this class -- do \b not compute or update the
residual/defect vector after each iteration! This is not an oversight, but it is a fundamental property of the
GMRES methods that both the solution vector as well as the residual/defect vectors are only computed after the
GMRES method has terminated or, in the case of restarted [F]GMRES(k) variants, after each restart, which happens
after at most k iterations. Saad describes this problem and proposes a solution in his book \cite Saad03 in
Proposition 6.9, which also states that the euclidean norm of the residual \f$\|b-Ax_m\|_2\f$ is equal to
a scalar quantity \f$|\gamma_{m+1}|\f$, which is easily computed from the other quantities that have to be
updated in each GMRES iteration. Therefore, one might simply use this scalar to implement a absolute or relative
defect tolerance stopping criterion without the need to explicitly compute the residual/defect vector itself.

However, in the real world case of finite precision arithmetic, the identity is merely an approximation, i.e.
we actually have \f$|\gamma_{m+1}|\approx\|b-Ax_m\|_2\f$ and therefore checking \f$|\gamma_{m+1}|\f$ against
the stopping criterion may lead to false positives and therefore premature termination of the inner FGMRES(k)
loop. At the end of each inner FGMRES(k) loop, the solution and residual/defect vectors are updated and now
the real residual/defect norm is computed and checked against the tolerances and it may now happen that
\f$|\gamma_{m+1}| < \textnormal{tol}_{\textnormal{abs}} < \|b-Ax_m\|_2\f\f$ (and analogously for the relative
tolerance, of course), therefore leading to another restart of the inner FGMRES(k) loop. In the worst case,
this restarted inner loop may again produce a false positive after 1 iteration, therefore triggering another
restart, thus effectively leading to an extremely slowly converging FGMRES(1) iteration.

To deal with this problem, this implementation includes an inner residual scaling factor \f$\delta\geq 0\f$,
which is used to check \f$\delta|\gamma_{m+1}| < \textnormal{tol}_{\textnormal{abs}}\f$ for the absolute and
\f$\delta|\gamma_{m+1}| < \textnormal{tol}_{\textnormal{rel}}\|r_0\|_2\f$ for the relative stopping criterion,
respectively. By choosing a \f$0 < \delta < 1\f$, one can ensure that the inner GMRES loop has to fulfill a
tighter tolerance than the outer loop, therefore reducing the risk of false positives. It is also possible to
set \f$\delta = 0\f$, in which case the inner GMRES loop will only terminate prematurely if it detects that it
has arrived at the exact solution and no more inner iterations must be performed to avoid division by zero.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>