BREAKING CHANGE: Major Refactor of build system
This commit refactors the way our build system works in a few
ways. First of all, library targets now use PUBLIC and PRIVATE linking
and always use target_... to set dependencies.
Project wide includes (e.g. include_directories, and so on) are now
banned and usage WILL be reported to authorities.
This does not change the way an executable consumes its dependencies,
so simply use target_link_libraries(...) and it should work fine.

The second major change is the way thirdparty dependencies are handled.
In contrast to before, now cmake itself handles download and linking of
TPLs, with the added benefit that you can now opt in into finding and using
externally installed thirdparty libraries.

In the end, when using the configure script, there is no change except
that the dependencies are no longer downloaded into the src folder,
but into the build folder. This will lead to a longer initial configure
step for each thirdparty activated, but this is only necessary once per
build folder. So this should be reasonable.
Be warned, there will be a lot more output in the configration step,
which can be mostly ignored.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>