changed LAFEM axpy interface from r.axpy(x, y, alpha=1) to y.axpy(x, alpha=1)
This commit introduces a breaking change in the axpy interface of all LAFEM
containers. Previously, the axpy function conformed to the interface

  r.axpy(x, y, alpha);

which performed the operation

  r <- y + alpha*x

where r, y and x may or may not be different objects.

The new interface of the LAFEM axpy functions introduced by this commit conforms
to the de-facto standard BLAS-1 style axpy operation of the form

  y <- y + alpha*x

which is the special case of the former implementation for r=y, and its
function call form looks like this

  y.axpy(x, alpha);

and the special case for alpha=1 which looks like this:

  y.axpy(x);

In consequence, all calls to LAFEM axpy functions have to be adjusted to the
new interface, because the code will not compile otherwise. Since the old
interface offered more options, it is necessary to replace old axpy calls by
their new counterparts, possibly in combination with a copy or scale call.

In principle, there are four interesting cases:

1) r = y

This is the default AXPY case and if your code calls

  y.axpy(x, y, alpha);

you can replace that call directly by

  y.axpy(x, alpha);

i.e. simply delete the second function call argument.

2) r = x, alpha = 1

If your code performs a simple vector addition of the form

  x.axpy(x, y);

you can simply replace that by

  x.axpy(y);

i.e. simply delete the first function call argument.

3) r = x, alpha != 1

This is a somewhat rare case that looks like this:

  x.axpy(x, y, alpha);

which has to be replaced by two function calls:

  x.scale(x, alpha);
  x.axpy(y);

4) r != x, r != y

This is the generic case where r, x and y are 3 different objects, so the call

  r.axpy(x, y, alpha);

needs to be replaced by two function calls:

  r.copy(y);
  r.axpy(x, alpha);

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>