Introduce new group based assembly kernels
This commit changes the actual assembly kernels used in the burgers
voxel assembly to use a warp/group base strategy, i.e. multiple
threads, organized as a group, assemble one element together.
This leads to considerable performance benefits for small mesh sizes
due to a larger exposed parallelity.
Also for the matrix assembly this leads to an overall increase of
at least factor 2 compared to the naive kernel.
In case of the defect assembly, this appears to be less than optimal
performance for large matrix sizes, so there is still some
work to do.
But for now, we keep this approach due to the potential in strong
scaling.

Note: For now, either the deformation nor the poisson kernel was
modified, but all relevant components are implemented to do this
analogeuos to the burgres assembler.

TODO: Test if a general domain assembler approach is significantly
slower.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>