added option to ignore local solver status codes to Global::SchwarzPrecond
The Global::SchwarzPrecond solver wrapper performs a global synchronization of
the status codes returned by the local solvers via an MPI allreduce operation
to ensure that the SchwarzPrecond returns a Status::success return code if and
only if all local solvers returned a successful return code. However, this
allreduce operation can become a bottleneck if the SchwarzPrecond is wrapping
a cheap and simple smoother (like AmaVanka) but is being called very frequently,
which is the default case in the CCND applications.

Therefore, the new option allows to ignore the local solver status codes, which
eliminates this bottleneck, especially since the local AmaVanka smoothers in
the CCND applications always return a successful status code anyways.

Please note that by default the local solver status codes are not ignored, since
in more sophisticated use cases of the SchwarzPrecond wrapper, e.g. in the
context of ScaRC-like solver configurations, the fairly expensive local solvers
may actually fail and the synchronization of these status codes is important
any they also should not become a bottleneck.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>