removed old-school #ifdef-#define-#endif header guards
Although #pragma once is by definition implementation dependent and therefore
not platform/compiler independent, we observe that practically all modern C++
compilers support this pragma and so it was high time to get rid of the
old-school header guards for good.

The only exception that is left over are the header files in kernel/lafem/arch,
because the guards also serve another purpose of checking proper include usage
there.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>