adjust copyright notice to include only year 2010
The first commit in the FEAT3 repository was pushed in 2010, so we decided that
the copyright notice of each source/header file should state this year in its
copyright notice, since it is a part of the project. We do not include the
latest year that the corresponding file was modified anymore, because this does
not seem to be that useful in continuous development cycle without classical
release versions anymore.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>