Added ostream flush function to communicator
This commit adds an explicit flush function, that simply calls
std::flush on the root rank.
Due to our changes regarding std::endl, it is now possible that all
printed statements are flushed to the standard output only after
the program has terminated. If the prorgamm is terminated irregulary,
for example by a sig kill triggered by the user, this could lead to lost
output.
Also, this can lead to different print timings depending on the used
system. Note that this behavior is intended.
To force printing at a specific time, you are now required to explicitly
flush with the new Comm::print_flush() function.

Also added print_flush in a few applications.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>