revisited Geometry::VoxelMap implementation to use vertex-based voxels
Previously, the Geometry::VoxelMap class and its domain control companion class
Control::Domain::VoxelDomainControl used cell-center-based voxels to discretize
a domain, which could be interpreted as a Q0 disretization, thus leading to the
name 'voxel' map. However, this led to the problem that a mesh created by the
VoxelDomainControl did not necessarily overlap the full actual domain, i.e.
some elements that barely intersected the discretized domain might get culled,
because the voxel centers were outside the domain while at least one voxel
corner might have been inside the domain.

Now, the internal implementation of the Geometry::VoxelMap is a vertex-based
discretization, which therefore can be interpreted as a Q1 discretization.
In consequence, the meshes generated by the VoxelDomainControl are now always
guaranteed to overlap the domain disretized by the voxel map, assuming that the
resolution of the voxel map is fine enough to capture all details of the
geometry.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>