standardcplusplusify Tiny algebra matrix Det/Vol/Inverse/Cofactor helpers
The previous implementations of the Det/Vol/Inverse/Cofactor/Helper classes
worked on raw DT_* pointers and performed the row/column indexing by pointer
arithmetic internally and the calling function simply had to pass the pointer
to the first element A(0,0) to the function, which worked just fine for over
a decade now. However, some overambitiously optimizing compilers, namely the
ICPX compiler, decided to ignore the pointer dependencies and therefore removed
a whole bunch of code that should not have been removed during the optimization
stage, therefore leading to completely corrupted applications, which failed due
to jacobian determinants erroneously being evaluted to zero.

I have changed the interface of the helper classes in the way that they now
accept the Tiny::Matrix objects themselves by reference and the internal helper
implementations now use the overloaded operator() to perform the indexing, thus
leading to text-book standard-conforming C++ code and the ICPX is happy now.

With that being said, my assessment is still that the previous implementation
was conforming to the C++17 standard, especially due to the rules in section
6.10 Lvalues and rvalues [basic.lval] and section 11.6.1 Aggregates [dcl.init.aggr]
which pretty much explicitly state that our use case does not fall under the
undefined behaviour commonly referred to as "type punning", i.e. in my opinion
it is the ICPX has overstepped its bounds provided by the C++ standards in this
case, but in the end it seemed safer to disarm this case before some other
compilers come up with similar ideas.

I have made some small benchmarks to check whether this change has negative
impacts on runtime performance and it seems that the changes in optimized build
are within the margin of error and therefore no slowdowns are to be expected.
I have not measured runtimes in debug mode, which I expected to rise due to an
excessively increased number of function calls, but that is not really relevant.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>