amputated Tiny::Vector/Matrix/Tensor3::size_cast() function
The size_cast() functions of the Tiny containers seemed like a good idea when I
wrote them, but it seems that they are deep in undefined behaviour land, and so
they have been removed. These functions have been used to interpret a Tiny
Vector of size n as a Tiny Vector of size m with m <= n, analogously for Matrix
and Tensor3 classes, which was useful for some occurances, but they have mostly
only been used because they exist and not because they provided any serious
performance or code elegance benefit.

To enable a similar functionality, this commit also adds a hand ful of new
functions to the Tiny containers, which allow to perform an operation on a
sub-block of the container, and these functions are suffixed by '_n'. The most
important function is the 'copy_n' function, which copes the first 'n' entries
of a source vector x of length nx >= n into the first 'n' entries a destination
vector y of length ny >= n and it is used as follows:

<code>
Tiny::Vector<DT_, 5> x(...);
Tiny::Vector<DT_, 3> y(...);
y.template copy_n<2>(x); // y[0] = x[0]; y[1] = x[1];
</code>

Other useful new '_n' functions are format_n, scale_n, axpy_n, normalize_n and
a few others that were added but are not being used until now. I did not provide
an '_n' version of every single Tiny container function, simply because they are
not required at the moment and they might be added in the future if the
necessity arises.

Most of the previous uses of size_cast() have been eliminated completely by
slightly refactoring the code and some other use cases have been eliminated by
making use of either copy_n<n>() or normalize_n<n>().

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>