added computation of domain coverage and volume to voxel map
The domain coverage is the ratio of active voxels to the total number of voxels
in the voxel map. The domain volume is the volume of the domain discretized by
the voxel map and it is defined as the product of the domain coverage and the
voxel map bounding box volume.

The domain coverage is also stored in the voxel map file header, therefore the
voxel map format has changed and old voxel maps are now incompatible, i.e.
existing voxel maps have to be recomputed.

Also replaced some #pragma omp ... statements by FEAT_PRAGMA_OMP(...)

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>