Add AdaptiveMesh class
The AdaptiveMesh class implements a new mesh data structure that allows
adaptively refining a ConformalMesh guided by subdivision levels
assigned to the vertices of the ConformalMesh.

The refinement is template based. That means mesh entities are replaced
with refined versions chosen from a fixed set of possible refinements.

The AdaptiveMesh supports:
* quadrilateral and hexahedral meshes,
* adaptation to different subdivision levels with minimal work
* exchangable template sets,
* a ConformalMesh-like interface for integrating with the rest of FEAT3,
* efficient retrieval of mesh entities, allowing assembly directly on
  the AdaptiveMesh.

The ConformalMesh-like interface has the following restrictions:

Left unsupported until absolutely required are
* VertexSet:begin() and VertexSet::end(),
* IndexSet::get_indices(),
which require changes to the APIs of VertexSet and IndexSet.

Left unsupported due to the work required is:
* ConformalMesh::deduct_topology_from_top(),
* ConformalMesh::reorient_boundary_facets().

ConformalMesh::reorient_boundary_facets() can be simulated by
reorienting the facets of the foundation mesh before adaptive
refinement.

Four different template sets are currently available:
* SchneidersTemplates
* SunZhaoMaTemplates
* SunZhaoMaExpansionTemplates
* TwoRefinementTemplates

Prefer the SchneidersTemplates for 2D refinement. The SunZhaoMa
templates produce less elements than the Schneiders templates in both 2D
and 3D, but produce worse adapter-elements, with decreasing angles for
multiple refinements. The SunZhaoMaExpansion templates are the only
templates that support refinemet of concave areas in 3D. The
TwoRefinementTemplates behave the same as the standard refinment of the
ConformalMesh.