Use identical host and system compiler
This change chooses as default the same host compiler for nvcc as the
chosen system compiler, e.g. if we configure with icx, icx is also
picked as nvcc host compiler.

To allow all compiler to be used, the option
--allow_unsupported_compiler
was added to our build config, which allows nvcc to use non officially
compilers. Use at your own risk.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>