Do not DeviceReset in runtime_finalize
Up until now, cuda_finalize() called cudaDeviceReset() which deallocates
all current device contexts of a single process.
Due to ordering problems with MPI_finalize() this leads to problems
with internal cuda resources of communication layers (e.g. UCX)
in the MPI backends.

Since calling DeviceReset by default is generally a bad idea, if
FEAT itself does not control the application flow, for example if
a FEAT application is simply called by another application or FEAT
is used as a backend in some other library, it is (by default)
deactivated.

Instead you can change the behavior by providing the
--finalize_resets_device configure flag or by defining
FEAT_FINALIZE_RESETS_DEVICE CMAKE variable.
If this option is set, cudaDeviceReset is called
as the very last call in the finalize method.

This could be necessary to have profilers and sanitizers work
correctly, but preliminary tests show this is not necessary
with recent cuda versions, so the default behavior appears to be
sufficient.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>