let PartiDomainControl ignore redundant level:process parameters
Previously, supplying a level:procs command line parameter like

  --level 7 5:16 3:16 2:1 0

to the PartiDomainControl has triggered an exception, because level 5 and 3 both
wanted to run on 16 processes, which was deemed illegal. From this commit on, a
level:procs parameter is simply ignored if the number of processes is identical
to the previous level:procs parameter, i.e above example is now treated as being
identical to

  --level 7 5:16 2:1 0

The primary reason for this change is that one can now always use something like

  --level N M:1 0

to go down to level 0 on a single process without requiring a special treatment
for the case where the application is started only with a single MPI process,
which would have triggered the above mentioned exception for the parameter 'M:1'
otherwise.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>