fixed a (mostly harmless) out-of-bounds access bug in Adjacency::Graph::render
This out-of-bounds access was triggered only when the last row of the adjactor
to be rendered was empty, which is a case that only appeared occasionally in
the test_matrix_factory-test, but not in real-world applications.

Also made said test write its RNG seed to cout, so that it is easier to
reproduce the test if it fails on once in 10000 runs or so.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>