use std::chrono::high_resolution_clock as Random seed
Previously, the Random class was either seeded by a default fixed seed or by
combining the output of time(nullptr) with Dist::Comm::world().rank(), the
latter of which is only defined once MPI has been initialized, which lead to
crashes in some cases.

The new high-resolution clock seed should be chaotic enough that different ranks
in a single MPI run should get different seeds without explicitly including the
MPI rank in the seed.

One major change is that the Random class is now seeded by the clock in the
default constructor, whereas the default constructor used the fixed seed until
now. As a consequence, mosts randomized unit tests now only use the default
constructor to seed the RNG by using the clock and the selected seed is now
always printed to cout in all unit tests that utilize the Random class.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>