fixed Global::Vector::[min/max]_element; assume gate ptr is always set
For async LA operations a gate is needed to return a SyncTicket.
Therefore, we now assume that for any (global) LA operation the gate
poiner always points to a valid Gate (this may be empty).

If such operations are now performed on a vetor that is not properly
initialized, the code will fail hard and early instead of running with
unsynchronized data which results in wrong results.

Instead this commit intreoducs asseriont (in debug mode) to check for
properly initialized gates.

Signed-off-by: Jonas Duennebacke <jduenneb@math.tu-dortmund.de>