added dedicated mesh-partitioner tool for partitioning pre-compuation
This new tool makes use of the ParMETIS and/or Zoltan third-party libraries to
read in a FEAT mesh file and apply one of the partitioning algorithms for a
to obtain a partitioning for a desired number of patches/partitions, which is
then written into a FEAT mesh file, so that it can be read in and used by the
Control::Domain::PartiDomainControl in a real-world application during startup.

This allows to separate or "out-source" the partitioning step from the actual
simulation step, which is useful to reduce the total runtime if several
simulations are to be performed on the same partitioning. Furthermore, this also
allows to execute the partitioner on a smaller number of MPI processes than the
actual simulation, which can also be useful for a large number of partitions.

Finally, this tool also allows a bit more configurability of the partitioning
step than the automatically performed partitioning in the PartiDomainControl
class, e.g. it is possible to specify whether the dual graph is to be set up by
facet adjacencies or by vertex adjacencies of the element.

In future, this tool may also be extended to give the caller even more control
over the parameters that are directly fed into the third-party libraries.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>