added OpenMP parallelization to Assembly::GridTransfer matrix assemblies
These do not use any sort of coloring or layering, but simply wrap the
potential race-condition in the scatter steps in a critical section.
Since the local mass matrix inversion takes up a lot of runtime, the hope
is that this will be sufficient to achieve good scalability without the
need for more sophisticated synchronization methods.

Unfortunately, I currently lack a proper profiling tool to verify this claim,
but this will hopefully change soon.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>