added OpenMP parallelization to Solver::AmaVanka numeric initialization
The numeric initialization of the AmaVanka smoother requires inverting fairly
large local matrices, so the scatter step was simply placed in a critical
section without any sort of coloring or layering. Whether this is sufficient
to achieve good scalability will have to be seen once I get my hands on a
proper parallel profiler.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>