added simple OpenMP parallelization to Assembly::DomainAssembler
The default multi-threading strategy is to use layering or coloring in combination
with a custom set of spawned worker threads, however, we also wanted to inlcude
simple OpenMP-based threaded parallelization to the DomainAssembler, even if
just to benchmark the threading strategies to figure out which ones work best.
As usual, this OpenMP-based parallelization makes use of a critical section
around the critical scatter step to avoid race conditions; this may become a
serial bottleneck for simple assemblies (e.g. of a vector), but it may not
be a huge problem for more complex assemblies like a 3D Burgers matrix.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>