implemented linear algebra backend parallel performance benchmark app
This benchmark measures the parallel performance of different linear algebra
backends by performing a modified power iteration loop on the 2D FEM 9-point
stencil matrix, which converges towards the eigenvector corresponding to the
largest eigenvalue of the matrix.
This benchmark application supports parallelization via MPI and OpenMP as well
as through the MKL and CUDA backends of the LAFEM containers. This benchmark
constructs the 9-point stencil matrix as well as the halo mirrors by hand, so
no actual mesh/trafo/space/assembly functionality is used by this benchmark
application to keep it simple.

The power iteration loop

  x_{k+1} := (1/|x_k|_2) * A * x_k

is split up into five sub-steps, which are measured separately:
1. compute euclidean norm of x_k locally on each process
2. perform an MPI allreduce to compute the global euclidean norm of x_k
3. scale x_k by the inverse of its global norm and save the result in vector t_k
4. compute local matrix-vector product x_k := A*_k
5. synchronize vector x_k by performing a halo exchange

This benchmark app measures the runtime of the five sub-steps separately and it
computes both the memory bandwidth as well as the floating point operations per
seconds of the three local sub-steps. Furthermore, this benchmark also includes
LIKWID markers for the three sub-steps that enable profiling on a hardware-
counter basis.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>