parallelized Adjacency::Graph render functions with OpenMP pragmas
Please note that the 'transpose' render functions have not been parallelized and
remain a purely sequential algorithm and I am unsure whether the transposition
could be parallelized efficiently anyways.

The good news is that the graph transposition makes up only a small portion of
the total symbolic assembly runtime, which is dominated by the rendering of the
composition and the subsequent index sorting.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>