Add cuda synchronization scope guard
Add new runtime member class and static variable controlling the
device synchronization behavior.
If the guard is in scope, all calls to cuda_synchronize() are
ommitted, in any other case, the bavior is the same.

Added a cuda_force_synchronization() function to circumvent
the new header.

Warning:
Only use this if you really know what you are doing.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>