Thirdparty CMake fixes
- Fixes version issue in CI pipelines

The manual find_package checks did not have minimum versions for
dependencies. This leads to the find_package(Boost ...) call succeeding
with Boost version 1.71, which is installed on the systems. CGAL then
later demands at least version 1.81, which is not available.

- Fixes Parmetis errors

Corrects a type in the FEAT_HAVE_PARMETIS option and unifies option
names between the thirdparty/CMakeLists.txt and the patched SuperLU
CMakeLists.txt.