Add caching of TPLs to CMake build system
Archives of thirdparty libraries are now downloaded to and kept in a
cache.

Use the --thirdparty_cache_dir flag of the configure script to set the
cache direcory. The default is the thirdparty/cache directory in the
FEAT repository that is being configured. Use the --no_thirdparty_cache
flag to disable caching behaviour. The CMake build system will then
always re-download the archives if required.

The cache directory is determined by the FEAT_TPL_CACHE_DIRECTORY
variable. If it is set, CMake will search for archives in the cache and
download missing archives to it. Unset (or set to an empty string) to
disable caching behaviour in CMake directly.

This commit also de-duplicates the TPL logic with a new helper function.