moved VS2022 kernel project file directly into kernel directory
The kernel project file for the Visual Studio 2022 build system has been
moved (back) to the kernel directory. The project was located in the
build_system/vc17 directory before, which required the user to create
a symbolic link to the kernel directory, which requires the user to have
admin rights on the machine, which may not be viable for all users.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>