major overhaul of the TraceAssembler class and trace assembly jobs/tasks
This commit transforms the previous provisional implementation of the TraceAssembler class
by an implementation that assembles jobs in a very similar fashion as the DomainAssembler.
In consequence, all of the TraceAssembler's previous assemble methods have been outsourced into
assembly task and job classes, which form an inheritance tree to outsource commonly used data
structures into common base classes, again in a very similar fashion to domain assembly jobs.

The TraceAssembler currently does not employ any sort of multi-threading, however, the job and task
classes are already designed to be executed in a multi-threaded fashion for upwards compatibility.

The functionality of the TraceAssembler is now split into three main files:

* kernel/assembly/trace_assembler.hpp: Contains the actual Assembly::TraceAssembler class.

* kernel/assembly/trace_assembler_basic_jobs.hpp: Contains the commonly used Task base classes as
  well as a large set of basic assembly jobs for the assembly of bilinearforms into matrices,
  linearforms into vectors as well as the integration of analytic and discrete functions.

* kernel/assembly/trace_assembler_jump_jobs.hpp: Contains a commonly used Task base class for
  the assembly of terms containing jump (or average) operators on the facets as well as two
  classes implementing the gradient-jump as well as the mass-jump operators which are used in the
  context of "edge-oriented jump stabilization" methods. These task classes could also be used
  for the assembly of discontinuous Galerkin operators in the future.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>