Convert VSCode support to CMake presets.
CMake presets[1] are the now recommended way to tell the VSCode
CMakeTools extension how to configure a project. This commit expands
the configure script to be able to create CMake presets. The configure
script can now take the role of the previous configureVscode.py script.

Whenever the configure script is called it will append a preset to the
CMakeUserPresets.json file in the project repo. Changes in this file are
automatically detected by VSCode. This allows VSCode to build FEAT in
exactly the same way the configure script would.

The configure script has been cleaned up. It now collects all compilers,
compiler flags, cmake cache entries, and other aspects of a FEAT
configuration in a FeatConfig object, which can then be used to generate
CMake presets or call out to CMake directly. The FeatConfig class also
checks for duplicate compiler choices and duplicate flags (turns out we
have been setting -Wrestrict and -Wnonnull twice this whole time).

Command line argument parsing has been replaced with the argparse
module. All flags have been standardized to the conventional
--foo-bar-baz format. The underscore versions are still available for
backwards compatability, but produce deprecation warnings. The
--ovr_mpi_ops flag has been renamed to --override-mpi-ops. Here too the
old version is still supported, but produces a warnings.

The behaviour of the configure script should be unchanged.

[1]: https://cmake.org/cmake/help/latest/manual/cmake-presets.7.html#id2