Added trace assembler for force free outflow
This commit introduces wrapper and trace assembly jobs
which allow to add the trace of the transposed gradient tensor
to the system matrix, allowing to transform the natural boundary
condition of the deformation formulation in the Stokes equation
to the one of the gradient formulation and vice versa.

For testing purposes, this was added to theccnd-dfg95 app and
can be activated by providing the --transform-outflow flag.

Signed-off-by: Maximilian Esser <maximilian.esser@math.tu-dortmund.de>