Expand ExportVTK class to allow export of mesh parts
This commit expands the ExportVTK class to allow exporting part of a
mesh or sub-manifolds of a mesh. As a side effect one can now also
export only the lower-dimensional entities of a mesh. For example, one
can export all faces of a hexahedral mesh as VTK quads, if desired.