Add boolean operations for MeshParts
This commits adds intersection, union, difference, and symmetric
difference operators for MeshParts.

The intersection and union operators support all mesh parts, with both
topologies and attributes. Topologies and attributes are copied from the
input mesh parts for all entities that are part of the final output.
For attributes a function has to be passed to determine how attributes
are handled if both input mesh parts assign a value to a vertex.

The difference and symmetric difference operators only support meshparts
without topologies for now. Attributes are still copied.