Add filtered constructors to all *HitTestFactory classes
The filtered constructors take an additional MeshPart parameter. Only
entities of that mesh-part are tested. All other entities are considered
not-hit. The result is thus always a subset of the filter mesh-part.

This commit also adds corresponding filtered versions of the
make_meshpart_by* factory functions and adds factory functions for the
ParsedHitTestFactory.